The conflict is a result of #11157 but is easy enough to resolve since

 Conflicts:
	src/api/InkAPI.cc

+++ src/tscore/MatcherUtils.cc
@@ -52,8 +52,8 @@ readIntoBuffer(const char *file_path, const char *module_name, int *read_size_pt
   int         fd;
   struct stat file_info;
   char       *file_buf, *buf;
-  int         read_size = 0;
-  int         file_size;
+  ssize_t     read_size = 0;
+  ssize_t     file_size;
 
   if (read_size_ptr != nullptr) {
     *read_size_ptr = 0;
@@ -113,7 +113,7 @@ readIntoBuffer(const char *file_path, const char *module_name, int *read_size_pt
     // Didn't get the whole file, drop everything. We don't want to return
     //   something partially read because, ie. with configs, the behaviour
     //   is undefined.
-    Error("%s Only able to read %d bytes out %d for %s file", module_name, read_size, file_size, file_path);
+    Error("%s Only able to read %ld bytes out %ld for %s file", module_name, read_size, file_size, file_path);
     ats_free(file_buf);
     file_buf = nullptr;
   }
